Include Some Readers

+++ src/PhpSpreadsheet/Reader/Csv/Delimiter.php
@@ -13,6 +13,7 @@ class Delimiter
 
     protected string $enclosure;
 
+    /** @var array<string, int[]> */
     protected array $counts = [];
 
     protected int $numberLines = 0;
@@ -53,6 +54,7 @@ class Delimiter
         }
     }
 
+    /** @param array<string, int> $delimiterKeys */
     protected function countDelimiterValues(string $line, array $delimiterKeys): void
     {
         $splitString = mb_str_split($line, 1, 'UTF-8');
@@ -69,7 +71,7 @@ class Delimiter
         // Calculate the mean square deviations for each delimiter
         //     (ignoring delimiters that haven't been found consistently)
         $meanSquareDeviations = [];
-        $middleIdx = floor(($this->numberLines - 1) / 2);
+        $middleIdx = (int) floor(($this->numberLines - 1) / 2);
 
         foreach (self::POTENTIAL_DELIMETERS as $delimiter) {
             $series = $this->counts[$delimiter];
